refactor(pipeline/scrape): extract EndpointResolver + reduce ScrapeAutoMapper to 27-eff shim
Phase 5 — Commit 8 of 9 (pipeline-decoupling-master-2026-05-28).

Final functional carve-out + shim conversion:

- EndpointResolver/EndpointResolver.ts (127 eff) takes the
  WK-bound TXN-endpoint resolution chain:
    * resolveAlias / resolveOptionalAlias — first-found WK
      alias against a sample record's keys.
    * pickAmountAlias — WK.amount → WK.creditAmount →
      WK.debitAmount fallback so the Beinleumi split-pair shape
      still passes the field-map check.
    * buildFieldMap — assembles the per-run ITxnFieldMap from
      the first record; returns false when neither date nor
      amount resolves (DASHBOARD.FINAL → F-DASH-2).
    * TEMPLATE_POST_LOOKUP / resolveTemplatePostData — replays
      the captured POST body verbatim on POST endpoints with a
      non-empty body, returns false otherwise.
    * resolvePendingUrl — `WK_API.pending` capture lookup with
      origin-based fallback to the canonical
      `/Transactions/api/approvals/getClearanceRequests` path.
    * billingBodyCarriesCardId / buildBillingUrlFromOrigin /
      resolveBillingUrl — billing URL discovery via
      WK_BILLING.pathFragment direct capture OR a
      WK_API.transactions capture whose POST body carries any
      WK_ACCT.queryId alias (per-card scoping marker).
    * EMPTY_FIELD_MAP / resolveFieldMapOrEmpty — empty-body
      pass-through for replayablePost tier captures so the
      URL + method are still committed and SCRAPE re-fetches per
      account.
    * resolveTxnEndpoint — the only public symbol; default
      export. Composes every helper into the slim
      ITxnEndpointInternal payload DASHBOARD.FINAL commits to
      `ctx.txnEndpoint`. Empty-object body coercion (Phase H')
      and method narrowing to 'GET' | 'POST' both preserved
      verbatim.

- ScrapeAutoMapper.ts is now a 27-eff PURE RE-EXPORT SHIM
  (38 raw, below the ≤40 LoC target). Every public symbol the
  pipeline + tests import from this path keeps working:
    * Re-exports from ScrapeReplayAction: IMonthChunk,
      buildMonthBody, generateMonthChunks, isMonthlyEndpoint,
      isRangeIterable, replaceField.
    * Re-exports from PipelineContext: TxnEndpoint, TxnFieldMap.
    * Re-exports from AccountExtractor: extractAccountIds,
      extractAccountRecords, extractAllContainers,
      isUsableIdentifier.
    * Re-exports from BfsFieldSearch: findAllFieldValues,
      findFieldValue, matchField.
    * Re-exports from Coercion: parseAutoDate.
    * Re-exports from ContainerPicker: extractTransactions,
      extractTransactionsForCard.
    * Re-exports from EndpointResolver: resolveTxnEndpoint
      (default → named).
    * Re-exports from FieldHunt/LifoCrawl: findFirstArray
      (default → named).
    * Re-exports from TxnMapper: autoMapTransaction.

Phase 5 split-end state (before Commit 9 ESLint guardrail):

  src/Scrapers/Pipeline/Mediator/Scrape/
  ├── AccountExtractor/
  │   ├── AccountExtractor.ts   (67 eff)
  │   ├── ContainerClaim.ts     (101 eff)
  │   └── RawShapeTrace.ts      (27 eff)
  ├── AutoMapperFacade/
  │   └── AutoMapperTypes.ts    (14 eff)
  ├── BfsFieldSearch/
  │   ├── BfsFieldSearch.ts     (138 eff)
  │   └── TxnSignature.ts       (23 eff)
  ├── Coercion/
  │   └── Coercion.ts           (29 eff)
  ├── ContainerPicker/
  │   └── ContainerPicker.ts    (56 eff)
  ├── EndpointResolver/
  │   └── EndpointResolver.ts   (127 eff)
  ├── FieldHunt/
  │   ├── LifoCrawl.ts          (106 eff)
  │   └── TxnHunt.ts            (60 eff)
  ├── TxnMapper/
  │   └── TxnMapper.ts          (96 eff)
  └── ScrapeAutoMapper.ts       (27 eff — pure shim)

Original ScrapeAutoMapper.ts: 1637 raw / 778 eff.
New shim:                     38 raw / 27 eff.

No production behaviour change. All 66 ScrapeAutoMapper-related
tests still pass unchanged. Husky 15-gate hook green.

Architecture test allowlist update (2 entries added — symmetric
file-path tracking for the split, no test-body or assertion
change): SCRAPE_DISCOVERY_ALLOWLIST and NET_SCRAPE_ALLOWLIST in
`LayerSeparationArchitecture.test.ts` now also include
`Mediator/Scrape/EndpointResolver/EndpointResolver.ts` since the
`discoverTransactionsEndpoint()` call + `INetworkDiscovery`
import moved with the resolveTxnEndpoint function. The original
`ScrapeAutoMapper.ts` entry stays (shim still re-exports the
function) so the invariant the test enforces is unchanged in
spirit — only the location of the privileged "parser" file is
updated.
